+++ main.tf
@@ -36,11 +36,11 @@ data "aws_iam_policy_document" "bucket_policy" {
     actions = [
       "s3:PutObject",
       "s3:GetObject",
-      "s3:GetObjectVersion"
+      "s3:GetObjectVersion",
     ]
     principals {
       type        = "AWS"
-      identifiers = formatlist("arn:aws:iam::%s:root", var.account_ids)
+      identifiers = formatlist("arn:aws:iam::%s:%s", var.account_ids, length(var.principals) == 0 ? [for x in var.account_ids : "root"] : var.principals)
     }
     resources = ["${aws_s3_bucket.bucket[0].arn}/*"]
   }
@@ -49,11 +49,11 @@ data "aws_iam_policy_document" "bucket_policy" {
     actions = [
       "s3:GetBucketLocation",
       "s3:ListBucket",
-      "s3:ListBucketVersions"
+      "s3:ListBucketVersions",
     ]
     principals {
       type        = "AWS"
-      identifiers = formatlist("arn:aws:iam::%s:root", var.account_ids)
+      identifiers = formatlist("arn:aws:iam::%s:%s", var.account_ids, length(var.principals) == 0 ? [for x in var.account_ids : "root"] : var.principals)
     }
     resources = [aws_s3_bucket.bucket[0].arn]
   }